% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{bh_asXYZ}
\alias{bh_asXYZ}
\title{XY list
Wrapper around rasterToPoint, produces a XYZ list.}
\usage{
bh_asXYZ(tissue = NULL)
}
\arguments{
\item{tissue}{tissue object}
}
\value{
array
}
\description{
XY list
Wrapper around rasterToPoint, produces a XYZ list.
}
