% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericShape.R
\name{bh_create}
\alias{bh_create}
\title{Object instantiation}
\usage{
bh_create(x, ...)
}
\arguments{
\item{x}{object of class genericShape or cellPrototype}

\item{lox}{numeric.}

\item{loy}{numeric, only for cell creation lox and loy are the coordinate
where to drop the cell}
}
\value{
a list containing an istantiation of a genericShape or an 
  an object of class cell.
}
\description{
The method bh_create cast a real shape or cell from a prototype.
}
