% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMarker.R
\name{bh_defineMarker}
\alias{bh_defineMarker}
\title{Define a new marker}
\usage{
bh_defineMarker(
  markerName = NULL,
  channelName = NULL,
  patternFunction = NULL,
  patternModifier = NULL,
  compartment = NULL
)
}
\arguments{
\item{markerName}{character, a name for the marker e.g. CD4.}

\item{channelName}{character, the channel (layer) in which the marker
will be detected. A corresponding layer should be created in a rasterStack
or tissue.}

\item{patternFunction}{character, an existing function, either one among
those coming in the package e.g. .pattern.skin() ore a custom function.}

\item{patternModifier}{named list, a list of parameters (e.g. mean, sd, tresh)
that specify new values. Parameters names must match those of the original
function. xy parameter is excluded because this parameter is passed 
automatically at the moment of taking a picture.}

\item{compartment}{character, one among cytoplasm, nucleus or organelle
(the last one not yet implemented), that specify the compartment that
express the marker.}
}
\value{
An object of class genericMarker
}
\description{
helper function to create a new marker.
}
