% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue.R
\name{bh_defineTissue}
\alias{bh_defineTissue}
\title{Define a new tissue}
\usage{
bh_defineTissue(coords = NULL, resolution = NULL, bg = NULL, markers = NULL)
}
\arguments{
\item{coords}{numeric vector, x and y limits... like (xMin,xMax,yMin,yMax).}

\item{resolution}{numeric vector, resolution on x and y.}

\item{bg}{numeric scalar, value to use in the background}

\item{markers}{list, makers.}
}
\value{
An object of class \strong{\link{tissue}}.
}
\description{
helper function to create a new tissue.
}
