% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate.R
\name{bh_loadPopulation}
\alias{bh_loadPopulation}
\title{Load a population.}
\usage{
bh_loadPopulation(file = NULL)
}
\arguments{
\item{file}{character scalar}
}
\description{
Just load a population saved to disk via \strong{\code{\link[=bh_savePopulation]{bh_savePopulation()}}}.
}
