% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate.R
\name{bh_populate_byClip}
\alias{bh_populate_byClip}
\title{Populate by clipping}
\usage{
bh_populate_byClip(
  cellPrototype = NULL,
  proportion = NULL,
  maxCloning = 10,
  tissue = NULL,
  cropToMesure = T,
  require_cytoplasm = T,
  require_nucleus = T,
  require_organelle = F,
  areaTresh = NULL
)
}
\arguments{
\item{cellPrototype}{list of objects of class cellPrototype}

\item{proportion}{numeric vector, proportion (0-1) of that specific cell}

\item{maxCloning}{numeric scalar, max of times the same cell can be cloned
around the area (may speed up the process).}

\item{tissue}{object of class \strong{\link{tissue}}}

\item{require_cytoplasm}{logical scalar, is cytoplasm otline required? cell
created without cytoplasm will be discarded. The same apply to require_nucleus
and require_organelle (not implemented at the moment).}

\item{require_nucleus}{logical scalar.}

\item{require_organelle}{logical scalar.}

\item{areaTresh}{numeric scalar, proportion (0-1) of tissue covered by cells.
Limit is calculated on the number of pixels that has either complete or
partial coverage.}

\item{cropToMeasure}{logical scalar, should cell outline cropped to the extent
of tissue?}
}
\value{
a list containing cells spatially distributed.
}
\description{
Populate comes in two versions, one is by clipping, meaning that cells are
formed without any knowledge of the surrounding cells and the overlaps will be
just clipped at the end. On the contrary, \strong{\code{\link[=bh_populate_byInteract]{bh_populate_byInteract()}}} creates
new cells considering neighboring cells.
}
