% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate.R
\name{bh_savePopulation}
\alias{bh_savePopulation}
\title{Save a population.}
\usage{
bh_savePopulation(x = NULL, file = NULL)
}
\arguments{
\item{x}{list of cells}

\item{file}{character scalar}
}
\description{
Just save a population created by either \strong{\link{bh_populate_byClip}} or
\strong{\link{bh_populate_byInteract}}
}
