% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{bh_saveSFC}
\alias{bh_saveSFC}
\title{XY list
Wrapper around st_sf, produces a data frame with geometries.}
\usage{
bh_saveSFC(sfc = NULL, filePath = NULL)
}
\arguments{
\item{sfc}{a sf object}

\item{filePath}{character, full path and file name. Extension .sqlite
will be added if missing.}
}
\value{
array
}
\description{
XY list
Wrapper around st_sf, produces a data frame with geometries.
}
