% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{bh_saveTiff}
\alias{bh_saveTiff}
\title{Save Tiff image
Wrapper around raster::writeRaster}
\usage{
bh_saveTiff(raster = NULL, filePath = NULL)
}
\arguments{
\item{raster}{a Raster Layer}

\item{filePath}{character, full path and file name. Extension .tiff
will be added if missing.}
}
\description{
Save Tiff image
Wrapper around raster::writeRaster
}
